Added all shapes to the component menu.
Added [DisallowMultipleComponent] attribute to shapes, rigid body and mass properties.

Fixed transform bug when RigidBody and Collide.[ShapeType] is on the same component level.

All debug rendering game objects are not serialized and recreated when the application changes states. Probably possible to add a flag (in the future) where the debug rendering isn't inclded in build.

Added AgXUnityEditor.Manager which loads with the Unity editor and manages behaviors and (in the future) tools/windows while working in the editor.