Taking Unity left handed coordinate system into account (ToHandedVec3, ToHandedVector3 etc. instead of AsVec3 etc. when transferring data).
Fixed so that shape positions are correct when scale is used. Collide.Mesh supports scale all the way (i.e., the size of the object is affected as well).

Fixed debug rendering of shapes to correctly match the geometries/shapes in the simulation.

Added button in simulation "Open in native viewer" to open the current scene in luaagx.