Improvements (#21)
* Added Constraint.GetController<T> to access the range, lock, motor etc of a constraint.

* Fixed bug where Shape.Material were set to null (by the rigid body instance) after each Property synchronization.

Improved GUI for bodies. All shapes are now displayed as a list in the body. It's also possible to drag & drop a ShapeMaterial to the rigid body GUI to assign material to all of its shapes.

* Manager is monitoring the number of shape debug rendering data components in a shape. If something doesn't seems right the debug render data is removed (will be created again if a DebugRenderManager is present).

* Added some VisualPrimitives.

Moved "create shape" code from AssemblyTool to a new tool named ShapeCreateTool.

Improved modelling of shapes from the AssemblyTool/ShapeCreateTool.

Bug fixes.

* Added SolveType to Constraint.

Improved tools.

* Added ConstraintCreateTool and DisableCollisionsTool, both possible to activate from RigidBodyTool. DisableCollisionsTool is present in ShapeTool and WireTool as well.

DisableCollisionsTool is using instanceId so it's possible to backtrack groups and which objects that are disabled. TODO.

* Collision group propagates to children if no RigidBody, Shape or Wire is in the selected game object.